Using XGBRFClassifier + Model Tuning
The business case

As a data specialist working for Salifort Motors, I have received the results of a recent employee survey. The senior leadership team has tasked me with analyzing the data to come up with ideas for how to increase employee retention. To help with this, they would like me to design a model that predicts whether an employee will leave the company based on their department, number of projects, average monthly hours, and any other data points I think are helpful.

The value of the deliverable

For this deliverable, I am asked to choose a method to approach this data challenge based on my prior work, selecting either a regression model or a machine learning model to predict whether an employee will leave the company.